Separate selection from extraction
