+++ selection/selection.py
@@ -1,51 +1,48 @@
 import os
+import random
 import shutil
 import glob
 from PIL import Image
 from osgeo import gdal
 from osgeo import ogr
 
-from extraction.extraction import select_random_map_images, copy_image_files
+from selection.errors import InvalidPathError, AbsolutePathError, OutputPathExistsError
+from selection.errors import InvalidTileSizeError, InsuffientDataError
 
 RASTER_TILE_SIZE = 10_000
 
 
-class InvalidPathError(Exception):
-    """Raised when the output path exists."""
-    pass
+class DataHandler():
+    def __init__(self):
+        pass
+
+    @staticmethod
+    def _verify_path(path_str):
+        if not os.path.exists(path_str):
+            raise InvalidPathError(f"Path {path_str} does not exist")
+        if os.path.isabs(path_str):
+            raise AbsolutePathError(f"Path {path_str} is absolute")
 
 
-class AbsolutePathError(Exception):
-    """Raised when the output path is absolute."""
-    pass
+class DataExtractor(DataHandler):
+    def __init__(self, input_path):
+        self._verify_path(input_path)
 
 
-class InvalidTileSizeError(Exception):
-    """Raised when the tile size is invalid."""
-    pass
-
-
-class InsuffientDataError(Exception):
-    """Raised when more data than available is requested."""
-    pass
-class DataSelector():
+class DataSelector(DataHandler):
 
     def __init__(self, input_path, testing: bool = False):
-
-        if not os.path.exists(input_path):
-            raise InvalidPathError(f"Input path {input_path} does not exist")
-
-        if os.path.isabs(input_path):
-            raise AbsolutePathError(f"Input path {input_path} is absolute")
-
+        self.extractor = DataExtractor(input_path)
         self.input_path = input_path
-        self.output_path = None
         self._testing = testing
 
     def select_data(self, tile_size, train_n, test_n, output_path, random_seed=0, lossy=False):
 
-        if os.path.isabs(output_path):
-            raise AbsolutePathError(f"Output path {output_path} is absolute")
+        # self.extractor.extract(tile_size=tile_size)
+
+        self._verify_path(output_path)
+        # if os.path.isabs(output_path):
+        #     raise AbsolutePathError(f"Output path {output_path} is absolute")
 
         self.output_path = os.path.join(
             output_path, f"selected_tiles_{tile_size}_{train_n}_{test_n}_{random_seed}")
@@ -76,21 +73,20 @@ class DataSelector():
         else:
             self._extract_data(tile_size, current_tile_path, train_n, test_n)
 
-        file_lists = select_random_map_images(
+        file_lists = self._select_random_map_images(
             train_size=train_n,
             test_size=test_n,
             input_path=current_tile_path,
             random_seed=random_seed,
         )
 
-        copy_image_files(
+        self._copy_image_files(
             file_lists,
             input_path=current_tile_path,
             output_path=self.output_path,
         )
 
     def _request_size_too_large(self, tile_size, train_n, test_n):
-
         if self._testing:
             # limit input to 32 tiles for faster testing
             raster_tile_size = min(self.raster_tile_size, tile_size*4)
@@ -105,8 +101,91 @@ class DataSelector():
 
         return total_no_of_pixels < requested_no_of_pixels
 
-    def _extract_data(self, tile_size, current_tile_path, train_n, test_n):
+    @staticmethod
+    def _select_random_map_images(train_size: int, test_size: int,
+                                  input_path: str, random_seed: int = 0) -> list:
+        """Selects a random subset of map images from the input path and returns
+        them as a List of two Lists (one each for train and test) of tuples pairs
+        of map and mask images.
 
+        Args:
+            train_size (int): number of training images to select
+            test_size (int): number of test images to select
+            input_path (str): location of input files
+
+        Returns:
+            list: List of lists of tuples pairs of map and mask images
+        """
+        # get all files in input directory
+        files = os.listdir(input_path)
+        files_map = [file for file in files if "map" in file]
+        files_mask = [
+            file for file in files if "msk" in file or "mask" in file]
+
+        # sort files by name
+        files_map.sort()
+        files_mask.sort()
+
+        files_zipped = list(zip(files_map, files_mask))
+
+        # shuffle the file pairs
+        random.seed(random_seed)
+        random.shuffle(files_zipped)
+
+        # select train and test from (shuffled) front
+        files_train = files_zipped[:train_size]
+        files_test = files_zipped[train_size:train_size+test_size]
+
+        # for now just return these as a list
+        # TODO define clearer file type for loaded data
+        return [files_train, files_test]
+
+    @staticmethod
+    def _copy_image_files(image_files: list, input_path: str,
+                          output_path: str, delete_existing_output_path_no_warning=False):
+        """Copy image files from the input path to the output path.
+
+        Args:
+            image_files (list): Filenames as returned by select_random_map_images
+            input_path (str): original file location
+            output_path (str): file location to copy to
+            delete_existing_output_path_no_warning (bool, optional): Delete output
+            path first if it exists, without warning. Defaults to False.
+        """
+
+        # if we have been asked to, delete existing out path without warning
+        # TODO is this safe?
+        if delete_existing_output_path_no_warning and os.path.exists(output_path):
+            shutil.rmtree(output_path)
+
+        # create output path if it doesn't exist or end if it does
+        if not os.path.exists(output_path):
+            os.makedirs(output_path)
+        else:
+            raise OutputPathExistsError("At least one of the output directory already exists."
+                                        "\nSet delete_existing=True to remove it.")
+
+        # get file names into a dict for easier processing
+        files = {}
+        files["train"] = image_files[0]
+        files["test"] = image_files[1]
+
+        for subfolder in ["train", "test"]:
+            output_path_subfolder = os.path.join(output_path, subfolder)
+
+            # create output folder if it doesn't exist
+            if not os.path.exists(output_path_subfolder):
+                os.makedirs(output_path_subfolder)
+
+            # copy files to output folder
+            for file_tuple in files[subfolder]:
+                for file_path in file_tuple:
+                    full_path_in = os.path.join(input_path, file_path)
+                    full_path_out = os.path.join(
+                        output_path_subfolder, file_path)
+                    shutil.copy(full_path_in, full_path_out)
+
+    def _extract_data(self, tile_size, current_tile_path, train_n, test_n):
         if self._testing:
             raster_tile_size = min(RASTER_TILE_SIZE, tile_size*4)
         else:
