+++ tests/test_data_extraction.py
@@ -1,4 +1,4 @@
-from extraction.extraction import select_random_map_images, copy_image_files
+from selection.selection import DataSelector
 import unittest
 import os
 import shutil
@@ -10,7 +10,7 @@ random.seed(42)
 
 
 class TestSelectMapImages(unittest.TestCase):
-    """Tests for select_random_map_images."""
+    """Tests for DataSelector._select_random_map_images."""
 
     def __init__(self, *args, **kwargs):
         """Initialise some shared test parameters."""
@@ -24,7 +24,7 @@ class TestSelectMapImages(unittest.TestCase):
 
     # TODO define saner data structure
     def test_select_map_images_returns_a_list_of_lists_of_tuples(self):
-        file_lists = select_random_map_images(
+        file_lists = DataSelector._select_random_map_images(
             train_size=1024,
             test_size=256,
             input_path=self.input_path,
@@ -36,7 +36,7 @@ class TestSelectMapImages(unittest.TestCase):
                 self.assertIsInstance(file_tuple, Tuple)
 
     def test_select_map_images_returns_two_file_lists(self):
-        file_lists = select_random_map_images(
+        file_lists = DataSelector._select_random_map_images(
             train_size=self.train_size,
             test_size=self.test_size,
             input_path=self.input_path,
@@ -45,7 +45,7 @@ class TestSelectMapImages(unittest.TestCase):
         self.assertEqual(len(file_lists), 2)
 
     def test_select_map_images_returns_correct_number_of_file_tuples_train(self):
-        file_lists = select_random_map_images(
+        file_lists = DataSelector._select_random_map_images(
             train_size=self.train_size,
             test_size=self.test_size,
             input_path=self.input_path,
@@ -53,7 +53,7 @@ class TestSelectMapImages(unittest.TestCase):
         self.assertEqual(len(file_lists[0]), self.train_size)
 
     def test_select_map_images_returns_correct_number_of_file_tuples_test(self):
-        file_lists = select_random_map_images(
+        file_lists = DataSelector._select_random_map_images(
             train_size=self.train_size,
             test_size=self.test_size,
             input_path=self.input_path,
@@ -61,7 +61,7 @@ class TestSelectMapImages(unittest.TestCase):
         self.assertEqual(len(file_lists[0]), self.train_size)
 
     def test_select_map_images_returns_two_existing_files_per_tuple(self):
-        file_lists = select_random_map_images(
+        file_lists = DataSelector._select_random_map_images(
             train_size=self.train_size,
             test_size=self.test_size,
             input_path=self.input_path,
@@ -78,7 +78,7 @@ class TestSelectMapImages(unittest.TestCase):
                     )
 
     def test_select_map_images_selects_only_tifs_train(self):
-        file_lists = select_random_map_images(
+        file_lists = DataSelector._select_random_map_images(
             train_size=self.train_size,
             test_size=self.test_size,
             input_path=self.input_path,
@@ -99,7 +99,7 @@ class TestSelectMapImages(unittest.TestCase):
 
 
 class TestCopyImageFiles(unittest.TestCase):
-    """Tests for copy_image_files."""
+    """Tests for DataSelector._copy_image_files."""
 
     def __init__(self, *args, **kwargs):
         """Initialise some shared test parameters."""
@@ -114,7 +114,7 @@ class TestCopyImageFiles(unittest.TestCase):
         """
         train_size = 10
         test_size = 5
-        image_files = select_random_map_images(
+        image_files = DataSelector._select_random_map_images(
             train_size=train_size,
             test_size=test_size,
             input_path=self.input_path
@@ -123,7 +123,7 @@ class TestCopyImageFiles(unittest.TestCase):
         if os.path.exists(self.output_path):
             shutil.rmtree(self.output_path)
 
-        copy_image_files(
+        DataSelector._copy_image_files(
             image_files,
             input_path=self.input_path,
             output_path=self.output_path,
